Refactor : dynamic import 세부적으로 적용
