+++ src/app/page.tsx
@@ -1,17 +1,16 @@
 import dynamic from 'next/dynamic'
 import Loading from './loading'
 
-const About = dynamic(async () => await import('@/components/about/About'), {
-  ssr: false,
-})
+const About = dynamic(async () => await import('@/components/about/About'))
 
-const Work = dynamic(async () => await import('@/components/work/Work'), {
-  ssr: false,
-})
+const Work = dynamic(async () => await import('@/components/work/Work'))
 
 const BusinessCardAnimation = dynamic(
   async () => await import('@/components/landing/BusinessCardAnimation'),
-  { ssr: false, loading: () => <Loading /> },
+  {
+    ssr: false,
+    loading: () => <Loading />,
+  },
 )
 
 export default function Home() {
