+++ src/components/@common/IntersectionObserverContainer.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React from 'react'
 import { useIntersectionObserver } from '@/lib/hooks/useIntersectionObserver'
