+++ src/components/about/EducationTimeline.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React from 'react'
 import styles from '@/styles/About.module.css'
 import ApexChart from 'react-apexcharts'
